requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.3
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
isort==5.8.0
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
matplotlib==3.4.1
mccabe==0.6.1
numpy==1.20.2
pandas==1.2.4
patsy==0.5.1
Pillow==8.2.0
pycodestyle==2.7.0
PyJWT==2.0.1
pylint==2.7.4
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
statsmodels==0.12.2
threadpoolctl==2.1.0
toml==0.10.2
ua-parser==0.10.0
url-parser==3.0.3
urllib3==1.26.4
virtualenv==20.4.3
wrapt==1.12.1
